+++ lib/dynflow/execution_plan/steps/abstract.rb
@@ -26,8 +26,8 @@ module Dynflow
         @error             = Type! error, ExecutionPlan::Steps::Error, NilClass
         @started_at        = Type! started_at, Time, NilClass
         @ended_at          = Type! ended_at, Time, NilClass
-        @execution_time    = Type! execution_time, Float
-        @real_time         = Type! real_time, Float
+        @execution_time    = Type! execution_time, Numeric
+        @real_time         = Type! real_time, Numeric
 
         self.state = state.to_sym
